create a local taxonomic database

create a local taxonomic database

td_create(authorities = "itis", schema = c("hierarchy", "taxonid",
  "synonyms"), overwrite = FALSE, lines = 1e+06,
  dbdir = rappdirs::user_data_dir("taxald"), db = td_connect(dbdir))

Arguments

authorities
a list (character vector) of authorities to be included in the database. By default, will install itis. See details for a list of recognized authorities. Use authorities="all" to install all available authorities automatically.

schema
format of the database to import.

overwrite
Should we overwrite existing tables? Default is FALSE. Change to "ask" for interactive interface, or TRUE to force overwrite (i.e. updating a local database upon new release.)

lines
number of lines that can be safely read in to memory at once. Leave at default or increase for faster importing if you have plenty of spare RAM.

dbdir
a location on your computer where the database should be installed. Defaults to user data directory given by rappdirs::user_data_dir.

db
connection to a database. By default, taxald will set up its own fast MonetDBLite::MonetDBLite connection.

Value

path where database has been installed (invisibly)

Details

Authorities recognized by taxald are:

itis

ncbi

col

tpl

gbif

fb FishBase, http://fishbase.org

slb, SeaLifeBase, http://sealifebase.org

wd, wikidata

Examples

# NOT RUN {
  # tmp <- tempdir()
  # create_db(authorities = "itis", dbdir = tmp)

# }